# api/docs/catalog-api.architecture.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ARIA Distribution Gateway — Catalog API",
    "description": "Browse, search, and install governed AI skills and agents from the ARIA catalog. AI platforms can use this API to discover capabilities and install them for end users.",
    "version": "1.0.0",
    "contact": {
      "name": "ARIA Platform Team",
      "email": "platform@aria.dev",
      "url": "https://aria.dev"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "ARIA Distribution Gateway"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/catalog/assets": {
      "get": {
        "operationId": "listAssets",
        "summary": "List or search available AI skills and agents",
        "description": "Returns catalog assets the caller is authorized to see. Supports free-text search and pagination.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "skill",
            "in": "query",
            "description": "Filter by OASF skill taxonomy (e.g., knowledge_retrieval/rag)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "domain",
            "in": "query",
            "description": "Filter by business domain (e.g., human_resources, engineering)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Free-text search across name, description, and tags",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "1-based page index",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          },
          {
            "name": "sensitivity",
            "in": "query",
            "description": "Filter by sensitivity tier",
            "schema": {
              "type": "string",
              "enum": [
                "public",
                "internal",
                "confidential",
                "highly_confidential"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of available assets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pageSize": {
                      "type": "integer"
                    },
                    "assets": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AssetListItem"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters"
          },
          "401": {
            "description": "Missing or invalid bearer token (when auth enforcement is enabled)"
          }
        }
      }
    },
    "/catalog/assets/{name}/{version}/manifest": {
      "get": {
        "operationId": "getAssetManifest",
        "summary": "Get full OASF record and governance details for an asset",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "description": "Asset name (URL-encoded), e.g., aria.dev%2Fskills%2Fhr-policy-lookup",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "description": "Semantic version, e.g., 1.2.0",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full asset manifest",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetManifest"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token (when auth enforcement is enabled)"
          },
          "402": {
            "description": "Asset invoke blocked: budget threshold exceeded"
          },
          "403": {
            "description": "Access denied by governance policy"
          },
          "404": {
            "description": "Asset not found"
          }
        }
      }
    },
    "/catalog/assets/{name}/{version}/install": {
      "post": {
        "operationId": "installAsset",
        "summary": "Submit an install request for an asset",
        "description": "Starts asynchronous install processing and returns a queued install identifier.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "target": {
                    "type": "string",
                    "enum": [
                      "claude_desktop",
                      "claude-desktop",
                      "vscode",
                      "cowork",
                      "web_portal",
                      "aria_cli"
                    ],
                    "default": "claude_desktop"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Install request accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "installId",
                    "status",
                    "estimatedReadyAt"
                  ],
                  "properties": {
                    "installId": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "accepted"
                      ]
                    },
                    "estimatedReadyAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "401": {
            "description": "Missing or invalid bearer token (when auth enforcement is enabled)"
          },
          "402": {
            "description": "Install blocked: asset budget threshold exceeded"
          },
          "403": {
            "description": "Install blocked by governance policy"
          },
          "404": {
            "description": "Asset not found"
          }
        }
      }
    },
    "/catalog/assets/{name}/{version}/request-access": {
      "post": {
        "operationId": "requestAssetAccess",
        "summary": "Request access to a governed asset",
        "description": "Submits an access request for an asset that the caller is not currently authorised to install.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "justification"
                ],
                "properties": {
                  "justification": {
                    "type": "string",
                    "description": "Business justification for access (min 8 characters)"
                  },
                  "ticketSystem": {
                    "type": "string",
                    "description": "Ticket system to use for the access request"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Access request submitted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "requestId",
                    "status",
                    "approvalChain"
                  ],
                  "properties": {
                    "requestId": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "submitted"
                      ]
                    },
                    "approvalChain": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (e.g. missing justification or caller already has access)"
          },
          "401": {
            "description": "Missing or invalid bearer token (when auth enforcement is enabled)"
          },
          "404": {
            "description": "Asset not found"
          }
        }
      }
    },
    "/catalog/cache/refresh": {
      "post": {
        "operationId": "refreshCatalogCache",
        "summary": "Force-refresh catalog metadata cache",
        "description": "Bypasses TTL and fetches the latest registry-backed catalog metadata. Use after publishing or removing assets.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "202": {
            "description": "Catalog cache refresh completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "stalenessWindowSeconds",
                    "freshnessSlaP95Seconds"
                  ],
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "refreshed"
                      ]
                    },
                    "refreshedAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    },
                    "stalenessWindowSeconds": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "freshnessSlaP95Seconds": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "p95FreshnessSeconds": {
                      "type": "number",
                      "nullable": true
                    },
                    "refreshSuccessTotal": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "refreshFailureTotal": {
                      "type": "integer",
                      "minimum": 0
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token (when auth enforcement is enabled)"
          },
          "502": {
            "description": "Catalog refresh failed due to registry request error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Entra (Azure AD) JWT bearer token. Obtain via OAuth 2.0 client-credentials or authorization-code flow against your tenant. Roles `aria-gateway-admin`, `aria-gateway-confidential`, and `aria-gateway-internal` control the sensitivity ceiling. Set AUTH_ENFORCE=true on the gateway to require this header."
      }
    },
    "schemas": {
      "AssetListItem": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Unique asset identifier"
          },
          "version": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "lifecycle_state": {
            "type": "string"
          },
          "sensitivity_tier": {
            "type": "string",
            "enum": [
              "public",
              "internal",
              "confidential",
              "highly_confidential"
            ]
          },
          "domains": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "trust_badge": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AssetManifest": {
        "type": "object",
        "properties": {
          "record": {
            "type": "object"
          },
          "governance": {
            "type": "object"
          },
          "install_url": {
            "type": "string"
          },
          "mcpb_url": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}